added error handling function
refactored code

added comment
