+++ TimeTrans.py
@@ -1,8 +1,16 @@
+"""
+Author: ldqLDQ, Diaplo
+功能： 与时间转换相关的操作
+"""
 import time
 
 
 def unix_time(dt):
-    if dt=="":
+    """
+    Author: Diaplo
+    功能：将时间字符串转换为Unix时间戳
+    """
+    if dt == "":
         return -1
     # 转换成时间数组
     timeArray = time.strptime(dt, "%Y-%m-%dT%H:%M")
@@ -12,7 +20,10 @@ def unix_time(dt):
 
 
 def normal_time(unixtime):
-    # 转换成normaltime
+    """
+    Author: Diaplo
+    功能：将Unix时间戳转换为时间字符串
+    """
     time_local = time.localtime(unixtime)
     # 转换成新的时间格式(2016年05月05日20:28)
     dt = time.strftime("%Y.%m.%d %H:%M", time_local)
@@ -20,6 +31,10 @@ def normal_time(unixtime):
 
 
 def finished_status(stime,etime):
+    """
+    Author: ldqLDQ
+    功能：获取当前自习状态
+    """
     nowtime = int(time.time())
     if nowtime < stime:
         return "未开始"
